refactor command checks
